+++ Banking/src/main/java/com/lcf/app/services/AddressServiceImpl.java
@@ -13,14 +13,14 @@ public class AddressServiceImpl implements AddressService{
 	private AddressDao addressDao;
 	
 	@Override
-	public long insertAddress(AddressDetails address) {
+	public int insertAddress(AddressDetails address) {
 		
-		long id= addressDao.insertAddress(address);
+		int id= addressDao.insertAddress(address);
 		return id;
 	}
 
 	@Override
-	public AddressDetails getAddressById(long addressId) {
+	public AddressDetails getAddressById(int addressId) {
 		AddressDetails address = addressDao.getAddressById(addressId);
 		return address;
 	}
